+++ src/cqterrain/swivel/__init__.py
@@ -1,2 +1,5 @@
 from .SwivelBase import SwivelBase
 from .SwivelTop import SwivelTop
+
+from .Turbine import Turbine
+from .Swivel import Swivel