+++ changes.md
@@ -1,9 +1,14 @@
 ## Main wip
 
 # 3.8.5
+* Updated README.md
+* Updated documentation.md
 * Added swivel package
  * Added SwivelBase
  * Added SwivelTop
+ * Added Turbine
+ * Added Swivel
+   * Added example swivel_plate
 
 # 3.8.4
 * greeble